backend-django/requirements.txt
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-extensions==3.0.8
django-rest-auth==0.9.5
djangorestframework==3.11.1
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
oauthlib==3.1.0
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.7
Pygments==2.6.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
traitlets==5.0.4
urllib3==1.25.10
wcwidth==0.2.5
